requirements.txt
python-telegram-bot==20.7
telethon==1.33.1
google-generativeai==0.3.1
firebase-admin==6.2.0
pymongo==4.6.1
python-dotenv==1.0.0
aiohttp==3.9.1
textblob==0.17.1
pytz==2023.3.post1